refactor: consolidate auth-related files under features/auth
