+++ src/components/navigation-auth.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@/components/ui/button';
-import { auth } from '@/lib/auth';
+import { auth } from '@/features/auth/lib/auth';
 import Link from 'next/link';
 import { UserMenu } from './user-menu';
